[refactor][jinj00oo]: 공통 컴포넌트와 회칙 전용 컴포넌트 분리 (#51)
