+++ src/features/bylaws/model/types.ts
@@ -1,0 +1,8 @@
+// 회칙 관련 데이터 구조
+// 아코디언 안에 들어가는 개별 아이템
+
+export type AccordionItemData = {
+  title: string;
+  scoreChange: string;
+  descriptions?: string[];
+};
